Refactor CLI; add ANYCID exploit support, BABE output, and PNX5230 backups
- Refactor and clean up argument parser to use action enum for faster parsing and clearer dispatch.

- Add global --anycid flag to execute flashed loader from Java exploit.

- Add read-flash option --save-as-babe to save dumps in BABE/SSW format.

- Implement ANYCID exploit support for DB2012, DB2020 and PNX5230, including restored boot area.

- Add firmware restore files to repository and security-units backup routine for PNX5230.

- Reduce serial read/write timeouts used when flashing firmware.

- Add: new loader for ANYCID

- Misc: formatting, help text updates, and better status logging.

+++ src/babe.h
@@ -54,8 +54,8 @@ enum
     COLOR_UNKNOWN = 0x4E4B4E55, //'UNKN' //?????????
 };
 
-int babe_check(uint8_t *file, uint32_t size, int checktype);
-int is_babe(uint8_t *addr, uint32_t size);
+int babe_check(uint8_t *file, size_t size, int checktype);
+int is_babe(uint8_t *addr, size_t size);
 
 typedef enum
 {
